refactor: move utils to common
