+++ recording-oracle/src/common/filters/exception.ts
@@ -9,7 +9,7 @@ import { Request, Response } from 'express';
 
 import { DatabaseError } from '@/common/errors/database';
 import logger from '@/logger';
-import { transformKeysFromCamelToSnake } from '@/utils/case-converter';
+import { transformKeysFromCamelToSnake } from '@/common/utils/case-converter';
 
 @Catch()
 export class ExceptionFilter implements IExceptionFilter {
