+++ recording-oracle/src/common/utils/environment.ts
