+++ recording-oracle/src/common/utils/type-guard.ts
