+++ recording-oracle/src/database/database.module.ts
@@ -8,7 +8,7 @@ import { RefreshTokenEntity } from '@/modules/auth';
 import { CampaignEntity, UserCampaignEntity } from '@/modules/campaigns';
 import { ExchangeApiKeyEntity } from '@/modules/exchange-api-keys';
 import { UserEntity } from '@/modules/users';
-import Environment from '@/utils/environment';
+import Environment from '@/common/utils/environment';
 
 @Module({
   imports: [
