+++ recording-oracle/src/modules/auth/auth.service.spec.ts
@@ -11,7 +11,7 @@ import { AuthConfigService } from '@/config';
 import logger from '@/logger';
 import { UsersRepository, UsersService } from '@/modules/users';
 import { generateUserEntity } from '@/modules/users/fixtures';
-import * as web3Utils from '@/utils/web3';
+import * as web3Utils from '@/common/utils/web3';
 import { generateES256Keys } from '~/test/fixtures/crypto';
 import { generateEthWallet } from '~/test/fixtures/web3';
