+++ recording-oracle/src/modules/auth/auth.service.ts
@@ -5,7 +5,7 @@ import { DEFAULT_NONCE } from '@/common/constants';
 import { AuthConfigService } from '@/config';
 import logger from '@/logger';
 import { UserEntity, UsersRepository, UsersService } from '@/modules/users';
-import * as web3Utils from '@/utils/web3';
+import * as web3Utils from '@/common/utils/web3';
 
 import { AuthError, AuthErrorMessage } from './auth.errors';
 import { RefreshTokenEntity } from './refresh-token.entity';
