+++ recording-oracle/src/modules/campaigns/campaigns.module.ts
@@ -4,7 +4,7 @@ import { ExchangeModule } from '@/modules/exchange';
 import { ExchangeApiKeysModule } from '@/modules/exchange-api-keys';
 import { StorageModule } from '@/modules/storage';
 import { Web3Module } from '@/modules/web3';
-import { PgAdvisoryLock } from '@/utils/pg-advisory-lock';
+import { PgAdvisoryLock } from '@/common/utils/pg-advisory-lock';
 
 import { CampaignsController } from './campaigns.controller';
 import { CampaignsRepository } from './campaigns.repository';
